+++ src/ewoksserver/app/routes/execution/models.py
@@ -14,7 +14,7 @@ class EwoksJobInfo(BaseModel):
     job_id: Union[str, int] = Field(title="Workflow execution job identifier")
 
 
-class EwoksEvent(BaseModel):
+class EwoksEvent_v1(BaseModel):
     host_name: str = Field(title="Host where the job was executed")
     process_id: int = Field(title="Process ID where the job was executed")
     user_name: str = Field(title="User name under which the job was executed")
@@ -62,11 +62,50 @@ class EwoksEventFilter(BaseModel):
     error: Optional[bool] = Field(title="Workflow execution failed", default=None)
 
 
-class EwoksEventList(BaseModel):
-    jobs: List[List[EwoksEvent]] = Field(
+class EwoksEventList_v1(BaseModel):
+    jobs: List[List[EwoksEvent_v1]] = Field(
         title="Workflow execution jobs grouped per job ID"
     )
 
 
 class EwoksWorkerList(BaseModel):
     workers: Optional[List[str]] = Field(title="Available workers for execution")
+
+
+class EwoksQueueList(BaseModel):
+    queues: Optional[List[str]] = Field(title="Available queues for execution")
+
+
+class EwoksEvent_v2(BaseModel):
+    host_name: str = Field(title="Host where the job was executed")
+    process_id: int = Field(title="Process ID where the job was executed")
+    user_name: str = Field(title="User name under which the job was executed")
+    job_id: str = Field(title="Workflow execution job identifier")
+    engine: Optional[str] = Field(title="Workflow execution engine", default=None)
+    context: str = Field(title="Event context (job, workflow, node)")
+    workflow_id: Optional[str] = Field(title="Workflow identifier", default=None)
+    node_id: Optional[str] = Field(title="Workflow node identifier", default=None)
+    task_id: Optional[str] = Field(title="Workflow task identifier", default=None)
+    type: str = Field(title="Event type (start, end, progress)")
+    time: str = Field(title="Event context send time")
+    error: Optional[bool] = Field(title="Workflow execution failed", default=None)
+    error_message: Optional[str] = Field(
+        title="Workflow execution error message", default=None
+    )
+    error_traceback: Optional[str] = Field(
+        title="Workflow execution error traceback", default=None
+    )
+    progress: Optional[int] = Field(title="Task progress in percentage", default=None)
+    task_uri: Optional[str] = Field(title="Workflow task output URI", default=None)
+    input_uris: Optional[List[Dict]] = Field(
+        title="Workflow task input URIs", default=None
+    )
+    output_uris: Optional[List[Dict]] = Field(
+        title="Workflow task output URIs", default=None
+    )
+
+
+class EwoksEventList_v2(BaseModel):
+    jobs: List[List[EwoksEvent_v2]] = Field(
+        title="Workflow execution jobs grouped per job ID"
+    )