+++ src/ewoksserver/app/routes/execution/router.py
@@ -22,11 +22,13 @@ from ...models import EwoksSchedulingType
 from . import models
 from . import events
 
+_base_execution_router = APIRouter()
 v1_0_0_router = APIRouter()
 v1_1_0_router = APIRouter()
+v2_0_0_router = APIRouter()
 
 
-@v1_0_0_router.post(
+@_base_execution_router.post(
     "/execute/{identifier}",
     summary="Execute workflow",
     response_model=models.EwoksJobInfo,
@@ -110,10 +112,13 @@ def execute_workflow(
     return {"job_id": future.task_id}
 
 
+v1_0_0_router.include_router(_base_execution_router)
+
+
 @v1_0_0_router.get(
     "/execution/events",
     summary="Get workflow events",
-    response_model=models.EwoksEventList,
+    response_model=models.EwoksEventList_v1,
     response_description="Workflow execution jobs grouped per job ID",
     status_code=200,
     responses={
@@ -123,7 +128,7 @@ def execute_workflow(
         },
     },
 )
-def execute_events(
+def execute_events_v1(
     settings: EwoksSettingsType,
     filters: Annotated[
         models.EwoksEventFilter, Depends(models.EwoksEventFilter)
@@ -146,6 +151,9 @@ def execute_events(
     return {"jobs": list(jobs.values())}
 
 
+v1_1_0_router.include_router(v1_0_0_router)
+
+
 @v1_1_0_router.get(
     "/execution/workers",
     summary="Get workers",
@@ -160,4 +168,52 @@ def workers(settings: EwoksSettingsType) -> Dict[str, Optional[List[str]]]:
     return {"workers": get_queues()}
 
 
-v1_1_0_router.include_router(v1_0_0_router)
+v2_0_0_router.include_router(_base_execution_router)
+
+
+@v2_0_0_router.get(
+    "/execution/events",
+    summary="Get workflow events",
+    response_model=models.EwoksEventList_v2,
+    response_description="Workflow execution jobs grouped per job ID",
+    status_code=200,
+    responses={
+        status.HTTP_503_SERVICE_UNAVAILABLE: {
+            "description": "Server not configured for ewoks events",
+            "model": common_models.ResourceIdentifierError,
+        },
+    },
+)
+def execute_events_v2(
+    settings: EwoksSettingsType,
+    filters: Annotated[
+        models.EwoksEventFilter, Depends(models.EwoksEventFilter)
+    ],  # pydantic model to parse query parameters
+) -> Dict[str, List[List[Dict]]]:
+    jobs = OrderedDict()
+    with events.reader_context(settings) as reader:
+        if reader is None:
+            raise HTTPException(
+                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
+                detail="Server not configured for ewoks events",
+            )
+        for event in reader.get_events(**filters.model_dump(exclude_none=True)):
+            job_id = event["job_id"]
+            if job_id not in jobs:
+                jobs[job_id] = list()
+            jobs[job_id].append(event)
+    return {"jobs": list(jobs.values())}
+
+
+@v2_0_0_router.get(
+    "/execution/queues",
+    summary="Get queues",
+    response_model=models.EwoksQueueList,
+    response_description="List of available queues",
+    status_code=200,
+)
+def queues(settings: EwoksSettingsType) -> Dict[str, Optional[List[str]]]:
+    if settings.ewoks_scheduling.type == EwoksSchedulingType.Local:
+        return {"queues": None}
+
+    return {"queues": get_queues()}