+++ src/ewoksserver/tests/test_discover.py
@@ -1,18 +1,18 @@
-from celery.exceptions import TimeoutError
 import pytest
-from .api_versions import ROOT_ALL_VERSIONS
+from celery.exceptions import TimeoutError
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_discover_tasks_from_a_module(rest_client, default_task_identifiers, root):
-    response = rest_client.get(f"{root}/tasks")
+def test_discover_tasks_from_a_module(rest_client, default_task_identifiers, api_root):
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     assert sorted(data["identifiers"]) == sorted(default_task_identifiers)
 
     module = "ewoksserver.tests.dummy_tasks"
 
-    response = rest_client.post(f"{root}/tasks/discover", json={"modules": [module]})
+    response = rest_client.post(
+        f"{api_root}/tasks/discover", json={"modules": [module]}
+    )
     data = response.json()
     assert response.status_code == 200, data
     expected = [
@@ -22,7 +22,7 @@ def test_discover_tasks_from_a_module(rest_client, default_task_identifiers, roo
     ]
     assert sorted(data["identifiers"]) == sorted(expected)
 
-    response = rest_client.get(f"{root}/tasks")
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     expected = default_task_identifiers + [
@@ -33,26 +33,25 @@ def test_discover_tasks_from_a_module(rest_client, default_task_identifiers, roo
     assert sorted(data["identifiers"]) == sorted(expected)
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_discover_method_task_type(rest_client, root):
+def test_discover_method_task_type(rest_client, api_root):
     module = "ewoksserver.tests.dummy_tasks"
 
     response = rest_client.post(
-        f"{root}/tasks/discover", json={"modules": [module], "task_type": "method"}
+        f"{api_root}/tasks/discover",
+        json={"modules": [module], "task_type": "method"},
     )
     data = response.json()
     assert response.status_code == 200
     assert sorted(data["identifiers"]) == ["ewoksserver.tests.dummy_tasks.my_task3"]
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_discover_all_tasks(rest_client, default_task_identifiers, root):
-    response = rest_client.get(f"{root}/tasks")
+def test_discover_all_tasks(rest_client, default_task_identifiers, api_root):
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     assert sorted(data["identifiers"]) == sorted(default_task_identifiers)
 
-    response = rest_client.post(f"{root}/tasks/discover")
+    response = rest_client.post(f"{api_root}/tasks/discover")
     data = response.json()
     assert response.status_code == 200, data
     expected = default_task_identifiers + [
@@ -65,29 +64,27 @@ def test_discover_all_tasks(rest_client, default_task_identifiers, root):
     ]
     assert set(expected) <= set(data["identifiers"])
 
-    response = rest_client.get(f"{root}/tasks")
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     assert set(expected) <= set(data["identifiers"])
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_discover_in_a_non_existing_module(rest_client, root):
+def test_discover_in_a_non_existing_module(rest_client, api_root):
     response = rest_client.post(
-        f"{root}/tasks/discover", json={"modules": ["not_a_module"]}
+        f"{api_root}/tasks/discover", json={"modules": ["not_a_module"]}
     )
     data = response.json()
     assert response.status_code == 404, data
     assert "No module named" in data["message"]
 
-    response = rest_client.post(f"{root}/tasks/discover")
+    response = rest_client.post(f"{api_root}/tasks/discover")
     data = response.json()
     assert response.status_code == 200, data
     assert data["identifiers"]
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_discover_timeout(celery_discover_timeout_client, root):
+def test_discover_timeout(celery_discover_timeout_client, api_root):
     rest_client, _ = celery_discover_timeout_client
     with pytest.raises(TimeoutError):
-        rest_client.post(f"{root}/tasks/discover")
+        rest_client.post(f"{api_root}/tasks/discover")