+++ src/ewoksserver/tests/test_tasks.py
@@ -1,13 +1,10 @@
 import json
-import pytest
-from .api_versions import ROOT_ALL_VERSIONS
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_single_task(rest_client, root):
+def test_single_task(rest_client, api_root):
     identifier = "myproject.tasks.Dummy"
 
-    response = rest_client.get(f"{root}/task/{identifier}")
+    response = rest_client.get(f"{api_root}/task/{identifier}")
     data = response.json()
     expected = {
         "message": "Task 'myproject.tasks.Dummy' is not found.",
@@ -22,7 +19,7 @@ def test_single_task(rest_client, root):
         "task_type": "class",
         "required_input_names": ["a"],
     }
-    response = rest_client.post(f"{root}/tasks", json=task1a)
+    response = rest_client.post(f"{api_root}/tasks", json=task1a)
     data = response.json()
     assert response.status_code == 200, data
     expected = {
@@ -32,7 +29,7 @@ def test_single_task(rest_client, root):
     }
     assert data == expected
 
-    response = rest_client.get(f"{root}/task/{identifier}")
+    response = rest_client.get(f"{api_root}/task/{identifier}")
     data = response.json()
     assert response.status_code == 200, data
     assert data == task1a
@@ -42,7 +39,7 @@ def test_single_task(rest_client, root):
         "task_type": "class",
         "required_input_names": ["a", "b"],
     }
-    response = rest_client.put(f"{root}/task/{identifier}", json=task1b)
+    response = rest_client.put(f"{api_root}/task/{identifier}", json=task1b)
     data = response.json()
     assert response.status_code == 200, data
     expected = {
@@ -52,17 +49,17 @@ def test_single_task(rest_client, root):
     }
     assert data == expected
 
-    response = rest_client.get(f"{root}/task/{identifier}")
+    response = rest_client.get(f"{api_root}/task/{identifier}")
     data = response.json()
     assert response.status_code == 200, data
     assert data == task1b
 
-    response = rest_client.delete(f"{root}/task/{identifier}")
+    response = rest_client.delete(f"{api_root}/task/{identifier}")
     data = response.json()
     assert response.status_code == 200
     assert data == {"identifier": identifier}
 
-    response = rest_client.delete(f"{root}/task/{identifier}")
+    response = rest_client.delete(f"{api_root}/task/{identifier}")
     data = response.json()
     assert response.status_code == 404
     assert data == {
@@ -71,7 +68,7 @@ def test_single_task(rest_client, root):
         "type": "task",
     }
 
-    response = rest_client.get(f"{root}/task/{identifier}")
+    response = rest_client.get(f"{api_root}/task/{identifier}")
     data = response.json()
     assert response.status_code == 404
     expected = {
@@ -82,29 +79,27 @@ def test_single_task(rest_client, root):
     assert data == expected
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_task_creation_errors(rest_client, default_task_identifiers, root):
+def test_task_creation_errors(rest_client, default_task_identifiers, api_root):
     task_without_id = {"task_type": "class"}
-    response = rest_client.post(f"{root}/tasks", json=task_without_id)
+    response = rest_client.post(f"{api_root}/tasks", json=task_without_id)
     assert response.status_code == 422
 
     task_with_empty_id = {"task_identifier": "", "task_type": "class"}
-    response = rest_client.post(f"{root}/tasks", json=task_with_empty_id)
+    response = rest_client.post(f"{api_root}/tasks", json=task_with_empty_id)
     assert response.status_code == 422
     data = response.json()
     assert data["message"] == "Task identifier cannot be empty"
 
     existing_id = default_task_identifiers[0]
     task_with_existing_id = {"task_identifier": existing_id, "task_type": "class"}
-    response = rest_client.post(f"{root}/tasks", json=task_with_existing_id)
+    response = rest_client.post(f"{api_root}/tasks", json=task_with_existing_id)
     assert response.status_code == 409
     data = response.json()
     assert data["message"] == f"Task '{existing_id}' already exists."
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_multiple_tasks(rest_client, default_task_identifiers, root):
-    response = rest_client.get(f"{root}/tasks")
+def test_multiple_tasks(rest_client, default_task_identifiers, api_root):
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     assert sorted(data["identifiers"]) == sorted(default_task_identifiers)
@@ -125,12 +120,12 @@ def test_multiple_tasks(rest_client, default_task_identifiers, root):
         "required_input_names": ["a", "b"],
     }
 
-    response = rest_client.post(f"{root}/tasks", json=task1a)
+    response = rest_client.post(f"{api_root}/tasks", json=task1a)
     data = response.json()
     assert response.status_code == 200, data
     assert data == task1a
 
-    response = rest_client.post(f"{root}/tasks", json=task1b)
+    response = rest_client.post(f"{api_root}/tasks", json=task1b)
     data = response.json()
     assert response.status_code == 409, data
     expected = {
@@ -140,12 +135,12 @@ def test_multiple_tasks(rest_client, default_task_identifiers, root):
     }
     assert data == expected
 
-    response = rest_client.post(f"{root}/tasks", json=task2)
+    response = rest_client.post(f"{api_root}/tasks", json=task2)
     data = response.json()
     assert response.status_code == 200, data
     assert data == task2
 
-    response = rest_client.get(f"{root}/tasks")
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200, data
     expected = default_task_identifiers + [
@@ -155,9 +150,8 @@ def test_multiple_tasks(rest_client, default_task_identifiers, root):
     assert sorted(data["identifiers"]) == sorted(expected)
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_task_descriptions(rest_client, default_task_identifiers, root):
-    response = rest_client.get(f"{root}/tasks/descriptions")
+def test_task_descriptions(rest_client, default_task_identifiers, api_root):
+    response = rest_client.get(f"{api_root}/tasks/descriptions")
     data = response.json()
     assert response.status_code == 200
     default_descriptions = [
@@ -169,11 +163,13 @@ def test_task_descriptions(rest_client, default_task_identifiers, root):
 
     module = "ewoksserver.tests.dummy_tasks"
 
-    response = rest_client.post(f"{root}/tasks/discover", json={"modules": [module]})
+    response = rest_client.post(
+        f"{api_root}/tasks/discover", json={"modules": [module]}
+    )
     data1 = response.json()
     assert response.status_code == 200, data1
 
-    response = rest_client.get(f"{root}/tasks/descriptions")
+    response = rest_client.get(f"{api_root}/tasks/descriptions")
     data2 = response.json()["items"]
     data2 = [
         r["task_identifier"] for r in data2 if r["task_identifier"].startswith(module)
@@ -182,8 +178,7 @@ def test_task_descriptions(rest_client, default_task_identifiers, root):
     assert sorted(data1["identifiers"]) == sorted(data2)
 
 
-@pytest.mark.parametrize("root", ROOT_ALL_VERSIONS)
-def test_malformed_task(rest_client, root, tmpdir):
+def test_malformed_task(rest_client, api_root, tmpdir):
     malformed_task_id = "myproject.tasks.Malformed"
     task_malformed = {"task_identifier": malformed_task_id}
     normal_task_id = "myproject.tasks.Normal"
@@ -191,10 +186,10 @@ def test_malformed_task(rest_client, root, tmpdir):
         "task_identifier": normal_task_id,
         "task_type": "class",
     }
-    response = rest_client.post(f"{root}/tasks", json=task_malformed)
+    response = rest_client.post(f"{api_root}/tasks", json=task_malformed)
     assert response.status_code == 422
 
-    response = rest_client.get(f"{root}/tasks/descriptions")
+    response = rest_client.get(f"{api_root}/tasks/descriptions")
     data = response.json()
     existing_tasks = len(data["items"])
 
@@ -207,15 +202,15 @@ def test_malformed_task(rest_client, root, tmpdir):
     with open(malformed_task_file, "w") as f:
         json.dump(task_malformed, f)
 
-    response = rest_client.get(f"{root}/tasks/{malformed_task_id}")
+    response = rest_client.get(f"{api_root}/tasks/{malformed_task_id}")
     assert response.status_code == 404
 
-    response = rest_client.get(f"{root}/tasks/descriptions")
+    response = rest_client.get(f"{api_root}/tasks/descriptions")
     data = response.json()
     assert response.status_code == 200
     assert len(data["items"]) - existing_tasks == 1
 
-    response = rest_client.get(f"{root}/tasks")
+    response = rest_client.get(f"{api_root}/tasks")
     data = response.json()
     assert response.status_code == 200
     assert len(data["identifiers"]) - existing_tasks == 1